Automatically locate MATLAB and MATLAB libraries on OS X and Linux
This patch allows MATLAB.jl to be used on most OS X and Linux systems
without setting environmental variables.

On Linux, the location of the MATLAB home directory is inferred from
the path to the matlab executable. On OS X, the Applications directory
is searched and the newest MATLAB version is used. The MATLAB_HOME
environmental variable may be used to override the default location.

Additionally, this patch provides the full path to libmx and libeng to
dlopen(). On the Linux and OS X systems I tested on, this eliminates
the need to set LD_LIBRARY_PATH or DYLD_FALLBACK_LIBRARY_PATH.