Registered trademark
[skip ci]

They also use TM (i.e not registered), but tat seems wrong in this case.

FYI: I thought conversely TM right for Python.jl, but that's for the logo only. There's a discussion about registering that package, pointing to this one as precedent (while also thought the name here not ideal...). ® is maybe not required for Python (in that case), but might be best to be on the safe side, here I think we really want o be right:

https://se.mathworks.com/company/aboutus/policies_statements/trademarks.html